improve overall stability + impl. view elements

+++ frontend/src/app/pages/marketplace/element-card/element-card.component.ts
@@ -9,7 +9,8 @@ import { ValuesPipe } from '../../../shared/pipes/values.pipe';
 import { GenericElementCardComponent } from '../generic-element-card/generic-element-card.component';
 import { VersionedElementModalComponent } from '../editor-modals/versioned-element-modal/versioned-element-modal.component';
 import { CollectionService } from '../../../core/exercise-element.service';
-import { EditingVersionedElementModalData } from '../editor-modals/base-versioned-element-submodal';
+import { EditingVersionedElementModalData, VersionedElementModalData } from '../editor-modals/base-versioned-element-submodal';
+import { ConfirmationModalService } from '../../../core/confirmation-modal/confirmation-modal.service';
 
 @Component({
     selector: 'app-element-card',
@@ -17,23 +18,17 @@ import { EditingVersionedElementModalData } from '../editor-modals/base-versione
     templateUrl: './element-card.component.html',
     styleUrl: './element-card.component.scss',
 })
-export class ElementCardComponent implements OnInit {
+export class ElementCardComponent {
     private readonly ngbModalService = inject(NgbModal);
     private readonly collectionService = inject(CollectionService);
+    private readonly confirmationService = inject(ConfirmationModalService)
 
     public readonly availableElements = input<ElementDto[]>([]);
 
-    public readonly collection = input<VersionedCollectionPartial>();
+    public readonly collection = input.required<VersionedCollectionPartial>();
     public readonly element = input.required<ElementDto>();
-    public readonly editable = input(true);
+    public readonly mode = input.required<EditingVersionedElementModalData<any>["mode"]>();
 
-    ngOnInit() {
-        if (this.editable() && this.collection() === undefined) {
-            throw new Error(
-                'ElementCardComponent is editable but no collection was provided.'
-            );
-        }
-    }
 
     openEditor() {
         const modal = this.ngbModalService.open(
@@ -43,7 +38,7 @@ export class ElementCardComponent implements OnInit {
             }
         );
         modal.componentInstance.data = {
-            mode: "edit",
+            mode: this.mode(),
             type: this.element().content.type,
             onSubmit: async (data, conflictResolution) => {
                 this.collectionService.updateElement(
@@ -60,6 +55,12 @@ export class ElementCardComponent implements OnInit {
     }
 
     async deleteElement() {
+        const confirmation = await this.confirmationService.confirm({
+            title: 'Element löschen',
+            description: `Möchten Sie das Element "${this.element().title}" wirklich löschen? Diese Aktion kann nicht rückgängig gemacht werden.`,
+            confirmationButtonText: "Unwiederuflich löschen",
+        })
+        if(!confirmation) return;
         await this.collectionService.deleteElement(
             this.element().entityId,
             this.collection()!.entityId
